# api.yaml
openapi: 3.0.0
info:
  title: SweetPotato Server API
  version: '1.0'
  description: |-
    Sonolusの基本APIを拡張する感じ。
    ユーザー認証はFirebaseAuthorizationを通してやる。
  contact:
    name: お窓
    url: 'https://discord.gg/KEfVkfC6Q9'
  termsOfService: 'https://discord.gg/KEfVkfC6Q9'
  license:
    name: MIT
    url: 'https://opensource.org/licenses/mit-license.php'
servers:
  - url: 'http://localhost:5000'
    description: Local
paths:
...
  '/levels/{levelName}':
    parameters:
      - schema:
          type: string
        name: levelName
        in: path
        required: true
    get:
      summary: Get level
      tags:
        - levels
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLevelResponse'
        '404':
          description: Not Found
      operationId: getLevel
      description: |-
        It returns specified level info
        It will raise 404 if the level is not registered in this server
      parameters:
        - $ref: '#/components/parameters/localization'
    post:
      summary: Add level
      operationId: addLevel
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Conflict
      description: 指定された譜面情報をサーバーに登録します
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Level'
        description: ''
      tags:
        - levels
      security:
        - Bearer: []
    patch:
      summary: Edit level
      operationId: editLevel
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      description: 指定されたlevelを編集します
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Level'
      tags:
        - levels
      security:
        - Bearer: []
...
  '/tests/{testId}/levels/{levelName}':
    parameters:
      - schema:
          type: string
        name: testId
        in: path
        required: true
      - schema:
          type: string
        name: levelName
        in: path
        required: true
    get:
      summary: Get testing level
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLevelResponse'
        '404':
          description: Not Found
      operationId: getLevelTest
      description: |-
        It returns specified level info
        It will raise 404 if the level is not registered in this server
      tags:
        - tests
...
  '/users/{userId}/levels/{levelName}':
    parameters:
      - schema:
          type: string
        name: userId
        in: path
        required: true
      - schema:
          type: string
        name: levelName
        in: path
        required: true
    get:
      summary: Get users level
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLevelResponse'
        '404':
          description: Not Found
      operationId: getUsersLevel
      description: |-
        It returns specified level info
        It will raise 404 if the level is not registered in this server
      tags:
        - users
...
components:
  schemas:
    Level:
      description: |-
        A level provides a list of entities and powered by scripted behavior in engine, with skin, effect, background and particle, to create gameplay experience for players / It defines level for actual user play. It includes all data to play one level.
        https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Level
      type: object
      x-examples:
        example-1:
          value:
            name: h0GeHOGE
            version: 1
            rating: 1
            engine: string
            title: string
            artists: string
            author: string
            cover: 'http://example.com/'
            bgm: 'http://example.com/'
            data: 'http://example.com/'
            genre: general
            public: false
            userId: string
            description: string
        example-2:
          value:
            name: h0GeHOGE
            version: 1
            rating: 1
            engine: string
            useSkin:
              useDefault: true
            useBackground:
              useDefault: true
            useEffect:
              useDefault: true
            useParticle:
              useDefault: true
            title:
              ja: string
            artists:
              ja: string
            author:
              ja: string
            cover:
              type: LevelCover
              hash: string
              url: string
            bgm:
              type: LevelBgm
              hash: string
              url: string
            data:
              type: LevelData
              hash: string
              url: string
            genre: general
            public: false
            userId: string
            notes: 1
            description:
              ja: string
      title: Level
      properties:
        name:
          type: string
          minLength: 1
          description: english and number only name for searching
          maxLength: 100
          pattern: '^[0-9a-zA-Z]+$'
          example: h0GeHOGE
        version:
          type: integer
          default: 1
          minimum: 1
          example: 1
          description: Reserved for future update. current default is 1.
          maximum: 100
        rating:
          type: integer
          description: Difficulty of the level
          minimum: 1
          maximum: 1000
        engine:
          type: string
          description: Just a name of engine (Sonolus-express finds the engine from db)
        useSkin:
          type: object
          properties:
            useDefault:
              type: boolean
              description: 'If true (recommended), default resource specified by the engine will be used.'
            item:
              $ref: '#/components/schemas/Skin'
        useBackground:
          type: object
          properties:
            useDefault:
              type: boolean
              description: 'If true (recommended), default resource specified by the engine will be used.'
            item:
              $ref: '#/components/schemas/Background'
        useEffect:
          type: object
          properties:
            useDefault:
              type: boolean
              description: 'If true (recommended), default resource specified by the engine will be used.'
            item:
              $ref: '#/components/schemas/Effect'
        useParticle:
          type: object
          properties:
            useDefault:
              type: boolean
              description: 'If true (recommended), default resource specified by the engine will be used.'
            item:
              $ref: '#/components/schemas/Particle'
        title:
          $ref: '#/components/schemas/LocalizationText'
        artists:
          $ref: '#/components/schemas/LocalizationText'
        author:
          $ref: '#/components/schemas/LocalizationText'
        cover:
          $ref: '#/components/schemas/SonolusResourceLocator'
        bgm:
          $ref: '#/components/schemas/SonolusResourceLocator'
        data:
          $ref: '#/components/schemas/SonolusResourceLocator'
        genre:
          type: string
          description: '独自要素: 楽曲のジャンル'
          example: general
          minLength: 1
          maxLength: 20
          default: general
          pattern: '^[A-Za-z]+'
        public:
          type: boolean
          default: false
          description: '独自要素: 楽曲が全体公開かどうか'
        userId:
          type: string
          description: '独自要素: 譜面作成者のユーザーID'
          minLength: 1
          maxLength: 100
        notes:
          type: integer
          description: '独自要素: 譜面内のノーツ数'
          minimum: 1
          maximum: 10000000
        createdTime:
          type: integer
          description: '独自要素: データを作成したエポックミリ秒(ソート用)'
          minimum: 0
        updatedTime:
          type: integer
          description: '独自要素: データを更新したエポックミリ秒(ソート用)'
          minimum: 0
        description:
          $ref: '#/components/schemas/LocalizationText'
        coverHash:
          type: string
          description: '独自要素: サムネのハッシュ値(DBを兼ねるため)'
          maxLength: 50
        dataHash:
          type: string
          description: '独自要素: データのハッシュ(DBを兼ねるため)'
          maxLength: 50
        bgmHash:
          type: string
          description: '独自要素: BGMのハッシュ(DBを兼ねるため)'
          maxLength: 50
        playCount:
          type: integer
          description: '独自要素: プレイされた回数'
          minimum: 0
          maximum: 9999999999
      required:
        - name
        - version
        - rating
        - engine
        - title
        - artists
        - author
        - cover
        - bgm
        - data
        - genre
        - description
    Skin:
      description: |-
        A skin provides graphic elements to levels / It defines graphics for specific level
        https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Skin
        https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Server
      type: object
      x-examples: {}
      title: Skin
      properties:
        name:
          type: string
          minLength: 1
          description: english and number only name for searching
          maxLength: 100
        version:
          type: integer
          default: 1
          minimum: 1
          example: 1
          description: Reserved for future update. current default is 1.
          maximum: 100
        title:
          type: string
          minLength: 1
          description: base title of this content
          maxLength: 100
        subtitle:
          type: string
          minLength: 1
          description: something footer(ex. featuring xyz) for this content
          maxLength: 100
        author:
          type: string
          minLength: 1
          description: author of this content
          maxLength: 50
        thumbnail:
          $ref: '#/components/schemas/SonolusResourceLocator'
        data:
          $ref: '#/components/schemas/SonolusResourceLocator'
        texture:
          $ref: '#/components/schemas/SonolusResourceLocator'
        createdTime:
          type: integer
          description: '独自要素: データを作成したエポックミリ秒(ソート用)'
          minimum: 0
        updatedTime:
          type: integer
          description: '独自要素: データを更新したエポックミリ秒(ソート用)'
          minimum: 0
        userId:
          type: string
          description: '独自要素: 譜面作成者のユーザーID'
          minLength: 1
          maxLength: 100
        description:
          type: string
          description: '独自要素: サイト内および譜面情報欄に表示される説明文'
          minLength: 1
          maxLength: 3000
    Background:
      description: |-
        A background provides background graphical element to levels / It defines background used for specific level
        https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Background
      type: object
      x-examples: {}
      title: Background
      properties:
        name:
          type: string
          minLength: 1
          description: english and number only name for searching
          maxLength: 50
        version:
          type: integer
          default: 1
          minimum: 1
          example: 1
          description: Reserved for future update. current default is 1.
          maximum: 100
        title:
          type: string
          minLength: 1
          description: base title of this content
          maxLength: 100
        subtitle:
          type: string
          minLength: 1
          description: something footer(ex. featuring xyz) for this content
          maxLength: 100
        author:
          type: string
          minLength: 1
          description: author of this content
          maxLength: 50
        thumbnail:
          $ref: '#/components/schemas/SonolusResourceLocator'
        data:
          $ref: '#/components/schemas/SonolusResourceLocator'
        image:
          $ref: '#/components/schemas/SonolusResourceLocator'
        createdTime:
          type: integer
          description: '独自要素: データを作成したエポックミリ秒(ソート用)'
          minimum: 0
        updatedTime:
          type: integer
          description: '独自要素: データを更新したエポックミリ秒(ソート用)'
          minimum: 0
        userId:
          type: string
          description: '独自要素: 譜面作成者のユーザーID'
          minLength: 1
          maxLength: 100
        description:
          type: string
          description: '独自要素: サイト内および譜面情報欄に表示される説明文'
          minLength: 1
          maxLength: 3000
    Effect:
      description: |-
        An effect provides audio elements to levels / It defines audio effect used for specific level
        https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Effect
      type: object
      x-examples: {}
      title: Effect
      properties:
        name:
          type: string
          minLength: 1
          description: english and number only name for searching
          maxLength: 100
        version:
          type: integer
          default: 1
          minimum: 1
          example: 1
          description: Reserved for future update. current default is 1.
          maximum: 100
        title:
          type: string
          minLength: 1
          description: base title of this content
          maxLength: 100
        subtitle:
          type: string
          minLength: 1
          description: something footer(ex. featuring xyz) for this content
          maxLength: 50
        author:
          type: string
          minLength: 1
          description: author of this content
        thumbnail:
          $ref: '#/components/schemas/SonolusResourceLocator'
        data:
          $ref: '#/components/schemas/SonolusResourceLocator'
        createdTime:
          type: integer
          description: '独自要素: データを作成したエポックミリ秒(ソート用)'
          minimum: 0
        updatedTime:
          type: integer
          description: '独自要素: データを更新したエポックミリ秒(ソート用)'
          minimum: 0
        userId:
          type: string
          description: '独自要素: 譜面作成者のユーザーID'
          minLength: 1
          maxLength: 100
        description:
          type: string
          description: '独自要素: サイト内および譜面情報欄に表示される説明文'
          minLength: 1
          maxLength: 3000
    Particle:
      description: |-
        A particle provides particle effect elements to levels / It defines particle effect used for specific level
        https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Particle
      type: object
      x-examples: {}
      title: Particle
      properties:
        name:
          type: string
          minLength: 1
          description: english and number only name for searching
          maxLength: 100
        version:
          type: integer
          default: 1
          minimum: 1
          example: 1
          description: Reserved for future update. current default is 1.
          maximum: 100
        title:
          type: string
          minLength: 1
          description: base title of this content
          maxLength: 100
        subtitle:
          type: string
          minLength: 1
          description: something footer(ex. featuring xyz) for this content
          maxLength: 100
        author:
          type: string
          minLength: 1
          description: author of this content
          maxLength: 50
        thumbnail:
          $ref: '#/components/schemas/SonolusResourceLocator'
        data:
          $ref: '#/components/schemas/SonolusResourceLocator'
        texture:
          $ref: '#/components/schemas/SonolusResourceLocator'
        userId:
          type: string
          description: '独自要素: 譜面作成者のユーザーID'
          minLength: 1
          maxLength: 100
        createdTime:
          type: integer
          description: '独自要素: データを作成したエポックミリ秒(ソート用)'
          minimum: 0
        updatedTime:
          type: integer
          description: '独自要素: データを更新したエポックミリ秒(ソート用)'
          minimum: 0
        description:
          type: string
          description: '独自要素: サイト内および譜面情報欄に表示される説明文'
          minLength: 1
          maxLength: 3000
...
    SonolusResourceLocator:
      description: |-
        Sonolus Resource Locator(SRL) is used in server endpoints in order for Sonolus to locate a resource.
        All resources must be gzip compressed.
        https://github.com/NonSpicyBurrito/sonolus-wiki/wiki/Resource-and-SRL
      type: object
      x-examples:
        example:
          type: EffectData
          hash: addb1eb6402aa0dcb0d65c5a8dfd3935f08e3d39
          url: /repository/EffectData/addb1eb6402aa0dcb0d65c5a8dfd3935f08e3d39
      title: SonolusResourceLocator
      properties:
        type:
          type: string
          minLength: 1
          enum:
            - SkinData
            - EngineThumbnail
            - EffectData
            - SkinTexture
            - ParticleData
            - EffectThumbnail
            - LevelBgm
            - BackgroundThumbnail
            - EngineConfiguration
            - LevelData
            - ParticleTexture
            - BackgroundData
            - EngineData
            - BackgroundImage
            - LevelCover
            - SkinThumbnail
            - ParticleThumbnail
          example: LevelData
          maxLength: 50
          description: When specify type together with hash uniquely identifies a resource.
        hash:
          type: string
          minLength: 1
          description: Hash is calculated using SHA1 on the resource file
          maxLength: 100
        url:
          type: string
          minLength: 1
          description: 'If url starts with character /, it will be treated as relative path to server address.'
          maxLength: 500
...
    GetLevelResponse:
      description: Response struct of getLevel
      type: object
      x-examples: {}
      properties:
        item:
          $ref: '#/components/schemas/Level'
        description:
          type: string
          minLength: 1
        recommended:
          type: array
          items:
            $ref: '#/components/schemas/Level'
      required:
        - item
        - description
        - recommended
      title: GetLevelResponse
...
    LocalizationText:
      title: LocalizationText
      type: object
      description: ''
      properties:
        cn:
          type: string
          description: Simplified Chinese
          minLength: 1
          maxLength: 1000
        tw:
          type: string
          description: TraditionalChinese
          minLength: 1
          maxLength: 1000
        ja:
          type: string
          description: Japanese
          minLength: 1
          maxLength: 1000
        en:
          type: string
          description: English
          minLength: 1
          maxLength: 1000
        ko:
          type: string
          description: Korean
          minLength: 1
          maxLength: 1000
        id:
          type: string
          description: Indonesian
          minLength: 1
          maxLength: 1000
        es:
          type: string
          description: Spanish
          minLength: 1
          maxLength: 1000
      required:
        - ja
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
  parameters:
    localization:
      name: localization
      in: query
      required: false
      schema:
        type: string
        example: en
        maxLength: 50
        minLength: 0
        default: ja
      description: It localizes response items if possible
      allowEmptyValue: true
...
security: []